Qwen2.5_VL/requirements.txt
accelerate==1.10.1
aiofiles==23.2.1
annotated-types==0.7.0
anyio==4.10.0
av==15.1.0
certifi==2025.8.3
charset-normalizer==3.4.3
click==8.2.1
debugpy==1.8.16
exceptiongroup==1.3.0
fastapi==0.116.1
ffmpy==0.6.1
filelock==3.13.1
fsspec==2024.6.1
gradio==5.4.0
gradio_client==1.4.2
h11==0.16.0
hf-xet==1.1.9
httpcore==1.0.9
httpx==0.28.1
huggingface-hub==0.34.4
idna==3.10
Jinja2==3.1.4
markdown-it-py==4.0.0
MarkupSafe==2.1.5
mdurl==0.1.2
mpmath==1.3.0
networkx==3.3
numpy==2.1.2
nvidia-cublas-cu12==12.1.3.1
nvidia-cuda-cupti-cu12==12.1.105
nvidia-cuda-nvrtc-cu12==12.1.105
nvidia-cuda-runtime-cu12==12.1.105
nvidia-cudnn-cu12==9.1.0.70
nvidia-cufft-cu12==11.0.2.54
nvidia-curand-cu12==10.3.2.106
nvidia-cusolver-cu12==11.4.5.107
nvidia-cusparse-cu12==12.1.0.106
nvidia-nccl-cu12==2.20.5
nvidia-nvjitlink-cu12==12.9.86
nvidia-nvtx-cu12==12.1.105
orjson==3.11.3
packaging==25.0
pandas==2.3.2
pillow==11.0.0
psutil==7.0.0
pydantic==2.11.7
pydantic_core==2.33.2
pydub==0.25.1
Pygments==2.19.2
python-dateutil==2.9.0.post0
python-multipart==0.0.12
pytz==2025.2
PyYAML==6.0.2
qwen-vl-utils==0.0.10
regex==2025.9.1
requests==2.32.5
rich==14.1.0
ruff==0.12.12
safehttpx==0.1.6
safetensors==0.6.2
semantic-version==2.10.0
shellingham==1.5.4
six==1.17.0
sniffio==1.3.1
starlette==0.47.3
sympy==1.13.3
tokenizers==0.22.0
tomlkit==0.12.0
torch==2.4.0+cu121
torchaudio==2.4.0+cu121
torchvision==0.19.0+cu121
tqdm==4.67.1
transformers @ git+https://github.com/huggingface/transformers@bb45d3631ec7026db04a77d33a52b31766372160
triton==3.0.0
typer==0.17.4
typing-inspection==0.4.1
typing_extensions==4.12.2
tzdata==2025.2
urllib3==2.5.0
uvicorn==0.35.0
websockets==12.0
